requirements.txt
torch
torchvision
numpy
grpcio
protobuf
